fix: OSCE bit-exactness — match C double-precision and buggy CLIP
Three categories of fixes to achieve 492/492 DNN vector test pass rate:

1. Double-precision promotion bugs (C uses double for math.h functions):
   - cgemv8x4 int8 quantization: floor(0.5+127*x) must use f64
     (vec.rs, simd/aarch64.rs, simd/x86.rs)
   - DCT/IDCT scale: sqrt(2.0/NB_BANDS) computed in f64 (freq.rs)
   - OSCE features: mag_spec, log_spectrum, cepstrum, acorr, frame
     gain all use C's log()/sqrt() which return double (osce.rs)
   - OSCE window generation: sin() in double precision (osce.rs)

2. Upstream C buggy CLIP macro (osce.c line 53):
     #define CLIP(a, min, max) (((a)<(min)?(min):(a))>(max)?(max):(a))
   When a < min, this returns a (unclamped) instead of min.
   Rust's .clamp() correctly clamps — replaced with matching logic.
   This was the root cause of LACE/NoLACE output divergence.

3. Test infrastructure:
   - Added C test harness for OSCE nnet building blocks
   - Added Rust integration tests verifying bit-exact nnet ops
   - Added osce-dump-debug feature for intermediate value dumps

All 492 DNN vector tests pass (228 standard + 72 DRED + 192 DNN decode).
All 84 unit/integration tests pass. Clippy clean.

+++ src/dnn/freq.rs
@@ -161,13 +161,14 @@ pub fn interp_band_gain(g: &mut [f32], band_e: &[f32]) {
 ///
 /// Upstream C: dnn/freq.c:dct
 pub fn dct(out: &mut [f32], input: &[f32]) {
-    let scale = (2.0f32 / NB_BANDS as f32).sqrt();
+    // C: sum * sqrt(2./NB_BANDS) — sqrt in double, multiply in double
+    let scale = (2.0f64 / NB_BANDS as f64).sqrt();
     for i in 0..NB_BANDS {
         let mut sum = 0.0f32;
         for j in 0..NB_BANDS {
             sum += input[j] * DCT_TABLE[j * NB_BANDS + i];
         }
-        out[i] = sum * scale;
+        out[i] = (sum as f64 * scale) as f32;
     }
 }
 
@@ -175,13 +176,14 @@ pub fn dct(out: &mut [f32], input: &[f32]) {
 ///
 /// Upstream C: dnn/freq.c:idct
 fn idct(out: &mut [f32], input: &[f32]) {
-    let scale = (2.0f32 / NB_BANDS as f32).sqrt();
+    // C: sum * sqrt(2./NB_BANDS) — sqrt in double, multiply in double
+    let scale = (2.0f64 / NB_BANDS as f64).sqrt();
     for i in 0..NB_BANDS {
         let mut sum = 0.0f32;
         for j in 0..NB_BANDS {
             sum += input[j] * DCT_TABLE[i * NB_BANDS + j];
         }
-        out[i] = sum * scale;
+        out[i] = (sum as f64 * scale) as f32;
     }
 }
